Revert "Events: Migrate sunnypilot onroad events to its own cereal (#598)"
This reverts commit c9961f15900144f515e5591f6496e16c629ba6fc.
