+++ cereal/services.py
@@ -78,7 +78,6 @@ _services: dict[str, tuple] = {
   "modelManagerSP": (False, 1., 1),
   "selfdriveStateSP": (True, 100., 10),
   "longitudinalPlanSP": (True, 20., 10),
-  "onroadEventsSP": (True, 1., 1),
 
   # debug
   "uiDebug": (True, 0., 1),
