+++ selfdrive/selfdrived/tests/test_alertmanager.py
@@ -1,9 +1,7 @@
 import random
 
-from openpilot.selfdrive.selfdrived.events import Alert, EVENTS
+from openpilot.selfdrive.selfdrived.events import Alert, EmptyAlert, EVENTS
 from openpilot.selfdrive.selfdrived.alertmanager import AlertManager
-
-from openpilot.sunnypilot.selfdrive.selfdrived.events_base import EmptyAlert
 
 
 class TestAlertManager:
