+++ sunnypilot/mads/state.py
@@ -29,17 +29,14 @@ from openpilot.selfdrive.selfdrived.events import ET, Events
 from openpilot.selfdrive.selfdrived.state import SOFT_DISABLE_TIME
 from openpilot.common.realtime import DT_CTRL
 
-from openpilot.sunnypilot.selfdrive.selfdrived.events import EventsSP
-
 State = custom.SelfdriveStateSP.ModularAssistiveDrivingSystem.ModularAssistiveDrivingSystemState
 EventName = log.OnroadEvent.EventName
-EventNameSP = custom.OnroadEventSP.EventName
 
 ACTIVE_STATES = (State.enabled, State.softDisabling, State.overriding)
 ENABLED_STATES = (State.paused, *ACTIVE_STATES)
 
-GEARS_ALLOW_PAUSED_SILENT = [EventNameSP.silentWrongGear, EventNameSP.silentReverseGear, EventNameSP.silentBrakeHold,
-                             EventNameSP.silentDoorOpen, EventNameSP.silentSeatbeltNotLatched, EventNameSP.silentParkBrake]
+GEARS_ALLOW_PAUSED_SILENT = [EventName.silentWrongGear, EventName.silentReverseGear, EventName.silentBrakeHold,
+                             EventName.silentDoorOpen, EventName.silentSeatbeltNotLatched, EventName.silentParkBrake]
 GEARS_ALLOW_PAUSED = [EventName.wrongGear, EventName.reverseGear, EventName.brakeHold,
                       EventName.doorOpen, EventName.seatbeltNotLatched, EventName.parkBrake,
                       *GEARS_ALLOW_PAUSED_SILENT]
@@ -52,57 +49,45 @@ class StateMachine:
 
     self.state = State.disabled
 
-    self._events = Events()
-    self._events_sp = EventsSP()
-
   def add_current_alert_types(self, alert_type):
     if not self.selfdrive.enabled:
       self.ss_state_machine.current_alert_types.append(alert_type)
 
-  def check_contains(self, event_type: str) -> bool:
-    return bool(self._events.contains(event_type) or self._events_sp.contains(event_type))
-
-  def check_contains_in_list(self, events_list: list[int]) -> bool:
-    return bool(self._events.contains_in_list(events_list) or self._events_sp.contains_in_list(events_list))
-
-  def update(self, events: Events, events_sp: EventsSP):
+  def update(self, events: Events):
     # soft disable timer and current alert types are from the state machine of openpilot
     # decrement the soft disable timer at every step, as it's reset on
     # entrance in SOFT_DISABLING state
 
-    self._events = events
-    self._events_sp = events_sp
-
     # ENABLED, SOFT DISABLING, PAUSED, OVERRIDING
     if self.state != State.disabled:
       # user and immediate disable always have priority in a non-disabled state
-      if self.check_contains(ET.USER_DISABLE):
-        if events_sp.has(EventNameSP.silentLkasDisable) or events_sp.has(EventNameSP.silentBrakeHold):
+      if events.contains(ET.USER_DISABLE):
+        if events.has(EventName.silentLkasDisable) or events.has(EventName.silentBrakeHold):
           self.state = State.paused
         else:
           self.state = State.disabled
         self.ss_state_machine.current_alert_types.append(ET.USER_DISABLE)
 
-      elif self.check_contains(ET.IMMEDIATE_DISABLE):
+      elif events.contains(ET.IMMEDIATE_DISABLE):
         self.state = State.disabled
         self.add_current_alert_types(ET.IMMEDIATE_DISABLE)
 
       else:
         # ENABLED
         if self.state == State.enabled:
-          if self.check_contains(ET.SOFT_DISABLE):
+          if events.contains(ET.SOFT_DISABLE):
             self.state = State.softDisabling
             if not self.selfdrive.enabled:
               self.ss_state_machine.soft_disable_timer = int(SOFT_DISABLE_TIME / DT_CTRL)
               self.ss_state_machine.current_alert_types.append(ET.SOFT_DISABLE)
 
-          elif self.check_contains(ET.OVERRIDE_LATERAL):
+          elif events.contains(ET.OVERRIDE_LATERAL):
             self.state = State.overriding
             self.add_current_alert_types(ET.OVERRIDE_LATERAL)
 
         # SOFT DISABLING
         elif self.state == State.softDisabling:
-          if not self.check_contains(ET.SOFT_DISABLE):
+          if not events.contains(ET.SOFT_DISABLE):
             # no more soft disabling condition, so go back to ENABLED
             self.state = State.enabled
 
@@ -114,12 +99,12 @@ class StateMachine:
 
         # PAUSED
         elif self.state == State.paused:
-          if self.check_contains(ET.ENABLE):
-            if self.check_contains(ET.NO_ENTRY):
+          if events.contains(ET.ENABLE):
+            if events.contains(ET.NO_ENTRY):
               self.add_current_alert_types(ET.NO_ENTRY)
 
             else:
-              if self.check_contains(ET.OVERRIDE_LATERAL):
+              if events.contains(ET.OVERRIDE_LATERAL):
                 self.state = State.overriding
               else:
                 self.state = State.enabled
@@ -127,26 +112,26 @@ class StateMachine:
 
         # OVERRIDING
         elif self.state == State.overriding:
-          if self.check_contains(ET.SOFT_DISABLE):
+          if events.contains(ET.SOFT_DISABLE):
             self.state = State.softDisabling
             if not self.selfdrive.enabled:
               self.ss_state_machine.soft_disable_timer = int(SOFT_DISABLE_TIME / DT_CTRL)
               self.ss_state_machine.current_alert_types.append(ET.SOFT_DISABLE)
-          elif not self.check_contains(ET.OVERRIDE_LATERAL):
+          elif not events.contains(ET.OVERRIDE_LATERAL):
             self.state = State.enabled
           else:
             self.ss_state_machine.current_alert_types += [ET.OVERRIDE_LATERAL]
 
     # DISABLED
     elif self.state == State.disabled:
-      if self.check_contains(ET.ENABLE):
-        if self.check_contains(ET.NO_ENTRY):
-          if self.check_contains_in_list(GEARS_ALLOW_PAUSED):
+      if events.contains(ET.ENABLE):
+        if events.contains(ET.NO_ENTRY):
+          if events.contains_in_list(GEARS_ALLOW_PAUSED):
             self.state = State.paused
           self.add_current_alert_types(ET.NO_ENTRY)
 
         else:
-          if self.check_contains(ET.OVERRIDE_LATERAL):
+          if events.contains(ET.OVERRIDE_LATERAL):
             self.state = State.overriding
           else:
             self.state = State.enabled
