+++ sunnypilot/mads/tests/test_mads_state_machine.py
@@ -27,14 +27,13 @@ Last updated: July 29, 2024
 import pytest
 from pytest_mock import MockerFixture
 
-from cereal import custom
+from cereal import log, custom
 from openpilot.common.realtime import DT_CTRL
 from openpilot.sunnypilot.mads.state import StateMachine, SOFT_DISABLE_TIME, GEARS_ALLOW_PAUSED
-from openpilot.selfdrive.selfdrived.events import ET, NormalPermanentAlert
-from openpilot.sunnypilot.selfdrive.selfdrived.events import EVENTS_SP
+from openpilot.selfdrive.selfdrived.events import Events, ET, EVENTS, NormalPermanentAlert
 
 State = custom.SelfdriveStateSP.ModularAssistiveDrivingSystem.ModularAssistiveDrivingSystemState
-EventNameSP = custom.OnroadEventSP.EventName
+EventName = log.OnroadEvent.EventName
 
 # The event types that maintain the current state
 MAINTAIN_STATES = {State.enabled: (None,), State.disabled: (None,), State.softDisabling: (ET.SOFT_DISABLE,),
@@ -48,7 +47,7 @@ def make_event(event_types):
   event = {}
   for ev in event_types:
     event[ev] = NormalPermanentAlert("alert")
-  EVENTS_SP[0] = event
+  EVENTS[0] = event
   return 0
 
 
@@ -63,93 +62,87 @@ class TestMADSStateMachine:
   @pytest.fixture(autouse=True)
   def setup_method(self, mocker: MockerFixture):
     self.mads = MockMADS(mocker)
+    self.events = Events()
     self.state_machine = StateMachine(self.mads)
-    self.events = self.state_machine._events
-    self.events_sp = self.state_machine._events_sp
     self.mads.selfdrive.state_machine.soft_disable_timer = int(SOFT_DISABLE_TIME / DT_CTRL)
-
-  def reset(self):
-    self.events.clear()
-    self.events_sp.clear()
-    self.state_machine.state = State.disabled
 
   def test_immediate_disable(self):
     for state in ALL_STATES:
       for et in MAINTAIN_STATES[state]:
-        self.events_sp.add(make_event([et, ET.IMMEDIATE_DISABLE]))
+        self.events.add(make_event([et, ET.IMMEDIATE_DISABLE]))
         self.state_machine.state = state
-        self.state_machine.update(self.events, self.events_sp)
+        self.state_machine.update(self.events)
         assert State.disabled == self.state_machine.state
-        self.reset()
+        self.events.clear()
 
   def test_user_disable(self):
     for state in ALL_STATES:
       for et in MAINTAIN_STATES[state]:
-        self.events_sp.add(make_event([et, ET.USER_DISABLE]))
+        self.events.add(make_event([et, ET.USER_DISABLE]))
         self.state_machine.state = state
-        self.state_machine.update(self.events, self.events_sp)
+        self.state_machine.update(self.events)
         assert State.disabled == self.state_machine.state
-        self.reset()
+        self.events.clear()
 
   def test_user_disable_to_paused(self):
-    paused_events = (EventNameSP.silentLkasDisable, EventNameSP.silentBrakeHold)
+    paused_events = (EventName.silentLkasDisable, EventName.silentBrakeHold)
     for state in ALL_STATES:
       for et in MAINTAIN_STATES[state]:
-        self.events_sp.add(make_event([et, ET.USER_DISABLE]))
+        self.events.add(make_event([et, ET.USER_DISABLE]))
         for en in paused_events:
-          self.events_sp.add(en)
+          self.events.add(en)
           self.state_machine.state = state
-          self.state_machine.update(self.events, self.events_sp)
-          final_state = State.paused if self.events_sp.has(en) and state != State.disabled else State.disabled
+          self.state_machine.update(self.events)
+          final_state = State.paused if self.events.has(en) and state != State.disabled else State.disabled
           assert self.state_machine.state == final_state
-          self.reset()
+          self.events.clear()
 
   def test_soft_disable(self):
     for state in ALL_STATES:
       if state == State.paused:  # paused considers USER_DISABLE instead
         continue
       for et in MAINTAIN_STATES[state]:
-        self.events_sp.add(make_event([et, ET.SOFT_DISABLE]))
+        self.events.add(make_event([et, ET.SOFT_DISABLE]))
         self.state_machine.state = state
-        self.state_machine.update(self.events, self.events_sp)
+        self.state_machine.update(self.events)
         assert self.state_machine.state == State.disabled if state == State.disabled else State.softDisabling
-        self.reset()
+        self.events.clear()
 
   def test_soft_disable_timer(self):
     self.state_machine.state = State.enabled
-    self.events_sp.add(make_event([ET.SOFT_DISABLE]))
-    self.state_machine.update(self.events, self.events_sp)
+    self.events.add(make_event([ET.SOFT_DISABLE]))
+    self.state_machine.update(self.events)
     for _ in range(int(SOFT_DISABLE_TIME / DT_CTRL)):
       assert self.state_machine.state == State.softDisabling
       self.mads.selfdrive.state_machine.soft_disable_timer -= 1
-      self.state_machine.update(self.events, self.events_sp)
+      self.state_machine.update(self.events)
 
     assert self.state_machine.state == State.disabled
 
   def test_no_entry(self):
     for et in ENABLE_EVENT_TYPES:
-      self.events_sp.add(make_event([ET.NO_ENTRY, et]))
-      if not self.state_machine.check_contains_in_list(GEARS_ALLOW_PAUSED):
-        self.state_machine.update(self.events, self.events_sp)
+      self.events.add(make_event([ET.NO_ENTRY, et]))
+      if not self.events.contains_in_list(GEARS_ALLOW_PAUSED):
+        self.state_machine.update(self.events)
         assert self.state_machine.state == State.disabled
-        self.reset()
+        self.events.clear()
 
   def test_no_entry_paused(self):
     self.state_machine.state = State.paused
-    self.events_sp.add(make_event([ET.NO_ENTRY]))
-    self.state_machine.update(self.events, self.events_sp)
+    self.events.add(make_event([ET.NO_ENTRY]))
+    self.state_machine.update(self.events)
     assert self.state_machine.state == State.paused
 
   def test_override_lateral(self):
     self.state_machine.state = State.enabled
-    self.events_sp.add(make_event([ET.OVERRIDE_LATERAL]))
-    self.state_machine.update(self.events, self.events_sp)
+    self.events.add(make_event([ET.OVERRIDE_LATERAL]))
+    self.state_machine.update(self.events)
     assert self.state_machine.state == State.overriding
 
   def test_paused_to_enabled(self):
     self.state_machine.state = State.paused
-    self.events_sp.add(make_event([ET.ENABLE]))
-    self.state_machine.update(self.events, self.events_sp)
+    self.events.add(make_event([ET.ENABLE]))
+    self.state_machine.update(self.events)
     assert self.state_machine.state == State.enabled
 
   def test_maintain_states(self):
@@ -157,7 +150,7 @@ class TestMADSStateMachine:
       for et in MAINTAIN_STATES[state]:
         self.state_machine.state = state
         if et is not None:
-          self.events_sp.add(make_event([et]))
-        self.state_machine.update(self.events, self.events_sp)
+          self.events.add(make_event([et]))
+        self.state_machine.update(self.events)
         assert self.state_machine.state == state
-        self.reset()
+        self.events.clear()
